Fix #80 (find PyMalloc libpython)
Search for `libpython` now finds libraries of the form `libpython3.5.so`
and `libpython3.5m.so` including the `m` suffix for PyMalloc Python as
per PEP 3149.

In the event that libraries are found with and without the suffix, the
one without is used, so there should be no change for existing
RTensorFlow users with working setups.