workaround testthat bug
'testthat' imports 'waldo' imports 'rematch2' imports 'tibble',
except that on github actions runner macOS-latest with r-release (4.1) tibble is not
installed as part of `install_local(dependencies=TRUE)`. This only presents on macOS with R 4.1.
Likely a transient bug not worth tracking down.

This refactor replaces `expect_that` with `expect`,
to avoid the need for tibble.